Provide Working Example & Switch to Functions
In order to drive out some of the features I wanted to see I've
created an example in `test/support` that shows how this can
be used to build out your dsl library.  In doing so I've also
discovered that several of the macros should really be helper
functions instead, and have been changed as such.